requirements.txt
streamlit
pdf2image
Pillow
opencv-python-headless
numpy
